Reorder files for roster store slice
