+++ src/state/roster/calculations/faction.ts
@@ -1,6 +1,6 @@
-import { Faction, Factions, FactionType } from "../../types/factions.ts";
-import { Warband } from "../../types/warband.ts";
-import { getSpecialArmyOption } from "./buiding/special-rules.ts";
+import { Faction, Factions, FactionType } from "../../../types/factions.ts";
+import { Warband } from "../../../types/warband.ts";
+import { getSpecialArmyOption } from "./hero-special-stuff.ts";
 
 export function getFactionType(warbands: Warband[]): FactionType | "" {
   if (warbands && warbands.length <= 0) {
