+++ src/utils/specialRules.js
@@ -1,5 +1,3 @@
-import hero_constraint_data from "../assets/data/hero_constraint_data.json";
-
 /* Specific functionality to support special rules or certain units, factions or option types. */
 
 export const handleMirkwoodRangers = (roster, alliance_level) => {
@@ -157,94 +155,4 @@ export const handleRivendellElrond = (newRoster, elrondRemoved) => {
     return newWarband;
   });
   return newRoster;
-};
-
-export const handleSpecialArmyOption = (newRoster, warbandNum) => {
-  /* If a Hero is removed that provided a special army option, that option much be turned off for all eligible units
-  in the current army roster, and points updated. */
-  let specialArmyOption =
-    hero_constraint_data[newRoster.warbands[warbandNum - 1].hero.model_id][0][
-      "special_army_option"
-    ];
-  newRoster.warbands = newRoster.warbands.map((warband) => {
-    let newWarband = { ...warband };
-    newWarband.units = newWarband.units.map((_unit) => {
-      let newUnit = { ..._unit };
-      if (newUnit.name != null) {
-        newUnit.options = newUnit.options.map((_option) => {
-          let newOption = { ..._option };
-          if (
-            newOption.type === "special_army_upgrade" &&
-            newOption.opt_quantity === 1 &&
-            newOption.option === specialArmyOption
-          ) {
-            newRoster["points"] = newRoster["points"] - newUnit["pointsTotal"];
-            newWarband["points"] =
-              newWarband["points"] - newUnit["pointsTotal"];
-            newUnit["pointsPerUnit"] =
-              newUnit["pointsPerUnit"] - newOption.points;
-            newUnit["pointsTotal"] =
-              newUnit["pointsPerUnit"] * newUnit["quantity"];
-            newWarband["points"] =
-              newWarband["points"] + newUnit["pointsTotal"];
-            newRoster["points"] = newRoster["points"] + newUnit["pointsTotal"];
-            newOption["opt_quantity"] = 0;
-          }
-          return newOption;
-        });
-      }
-      return newUnit;
-    });
-    return newWarband;
-  });
-  return newRoster;
-};
-
-export const handleSpecialWarbandOption = (newRoster, warbandNum) => {
-  /* If a Hero is removed that provided a special warband option, that option much be turned off for all eligible units
-  in the current army roster, and points updated. */
-  newRoster.warbands = newRoster.warbands.map((warband) => {
-    let newWarband = { ...warband };
-    if (newWarband["num"] === warbandNum) {
-      newWarband.units = newWarband.units.map((_unit) => {
-        let newUnit = { ..._unit };
-        if (newUnit.name != null) {
-          newUnit.options = newUnit.options.map((_option) => {
-            let newOption = { ..._option };
-            if (
-              newOption.type === "special_warband_upgrade" &&
-              newOption.opt_quantity === 1
-            ) {
-              newRoster["points"] =
-                newRoster["points"] - newUnit["pointsTotal"];
-              newWarband["points"] =
-                newWarband["points"] - newUnit["pointsTotal"];
-              newUnit["pointsPerUnit"] =
-                newUnit["pointsPerUnit"] - newOption.points;
-              newUnit["pointsTotal"] =
-                newUnit["pointsPerUnit"] * newUnit["quantity"];
-              newWarband["points"] =
-                newWarband["points"] + newUnit["pointsTotal"];
-              newRoster["points"] =
-                newRoster["points"] + newUnit["pointsTotal"];
-              newOption["opt_quantity"] = 0;
-            }
-            return newOption;
-          });
-        }
-        return newUnit;
-      });
-    }
-    return newWarband;
-  });
-  return newRoster;
-};
-
-export const handle50PctBowLimit = (faction_data, alliance_level) => {
-  let new_faction_data = { ...faction_data };
-  new_faction_data["The Serpent Horde"].bow_limit =
-    alliance_level === "Historical" ? 0.5 : 0.33;
-  new_faction_data["Azog's Hunters"].bow_limit =
-    alliance_level === "Historical" ? 0.5 : 0.33;
-  return new_faction_data;
 };
